requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
gitpython>=3.1.40
pydantic>=2.0.0
python-multipart>=0.0.9
aiofiles>=23.0.0
anthropic>=0.40.0
tree-sitter==0.25.2
tree-sitter-python>=0.25.0
tree-sitter-javascript>=0.25.0
tree-sitter-typescript>=0.23.0
tree-sitter-go>=0.25.0
tree-sitter-rust>=0.24.0
tree-sitter-java>=0.23.0
